# Optimize Encode.