**This requires PR #11**.

This updates the `Encode` function to reduce the memory allocations needed to perform the encode and significantly speeds up performance for typical usage patterns.  It also improves the readability of the function a bit while here.

The following is a before and after comparison of a typical encode:

```
benchmark               old ns/op    new ns/op    delta
---------------------------------------------------------
BenchmarkBase58Encode   4411         1494         -66.13%

benchmark               old allocs   new allocs   delta
---------------------------------------------------------
BenchmarkBase58Encode   46           2            -95.65%

benchmark               old bytes    new bytes    delta
---------------------------------------------------------
BenchmarkBase58Encode   592          96           -83.78%
```